refactor to make each command a seperate file
