+++ function_discord_bot/discord_command_validate.py
@@ -1,0 +1,4 @@
+from flask.wrappers import Request
+
+def validate(request: Request):
+    return "looks legit to me. (note this is a dummy response)"
